some requested changes but not all - still stuff to do
