+++ components/Login/SignUp.tsx
@@ -6,34 +6,14 @@ import * as Yup from 'yup'
 import styled from 'styled-components'
 import Link from 'next/link'
 import { PasswordInputBox } from '../FormStyles/PasswordInputBox'
-
-export const Container = styled.div`
-  background-color: #ffffff;
-  display: flex;
-  justify-content: flex-start;
-  flex-flow: column;
-  gap: 20px;
-  align-items: flex-start;
-`
-
-export const Title = styled.h2`
-  font-family: Source Sans Pro;
-  font-style: normal;
-  font-weight: 600;
-  font-size: 34px;
-  line-height: 35px;
-`
-
-export const SubTitle = styled.p`
-  margin-bottom: 30px;
-`
-export const SubContainer = styled.div`
-  display: flex;
-  flex-flow: column;
-  align-items: flex-end;
-  justify-content: space-evenly;
-  gap: 20px;
-`
+import {
+  BackButton,
+  EStyledButton,
+  Container,
+  SubContainer,
+  Title,
+  SubTitle,
+} from './LoginStyling'
 
 export const PasswordDiv = styled.div`
   display: flex;
@@ -42,37 +22,15 @@ export const PasswordDiv = styled.div`
   align-items: flex-start;
   gap: 20px;
 `
-export const EmailDiv = styled.div``
 
-export const EStyledButton = styled(Button)`
-  margin-top: 20px;
-  align-self: flex-end;
-  :hover {
-    box-shadow: 0 12px 16px 0 rgba(0, 0, 0, 0.24),
-      0 17px 50px 0 rgba(0, 0, 0, 0.19);
-  }
-`
-
-export const BackButton = styled.button`
-  border: none;
-  background: none;
-  font-family: Source Sans Pro;
-  font-style: normal;
-  font-weight: 600;
-  font-size: 28px;
-  color: #5064c7;
-  margin-bottom: 30px;
-
-  :hover {
-    text-shadow: 2px 2px 4px gray;
-  }
-`
 interface FormValues {
   email: string
   password: string
   confirmPass: string
 }
 
+// component for signup page - includes form and validation for email and password,
+// and ensures that passwords are the same
 function Signup() {
   const initialVal: FormValues = {
     email: '',
@@ -104,7 +62,7 @@ function Signup() {
           </Link>
           <Title>Welcome</Title>
           <SubTitle>Let us get your account up and running.</SubTitle>
-          <EmailDiv>
+          <div>
             <StyledTextInput
               width="330px"
               height="53px"
@@ -112,7 +70,7 @@ function Signup() {
               placeholder="Email"
               type="text"
             />
-          </EmailDiv>
+          </div>
           <SubContainer>
             <PasswordDiv>
               <PasswordInputBox
